separates dynamicsurface, implements zoom tweening
